refactor: moved runESLint into utils.ts and removed runners and its corresponding file
